+++ packages/MJCoreEntitiesServer/src/custom/MJDuplicateRunEntityServer.server.ts
@@ -1,0 +1,25 @@
+import { BaseEntity, PotentialDuplicateRequest } from "@memberjunction/core";
+import { RegisterClass } from "@memberjunction/global";
+import { MJDuplicateRunEntity } from "@memberjunction/core-entities";
+import { DuplicateRecordDetector } from "@memberjunction/ai-vector-dupe";
+
+@RegisterClass(BaseEntity, 'MJ: Duplicate Runs')
+export class MJDuplicateRunEntityServer extends MJDuplicateRunEntity  {
+    public async Save(): Promise<boolean> {
+        const saveResult: boolean = await super.Save();
+        if (saveResult && this.EndedAt === null) {
+            // do something
+            const duplicateRecordDetector: DuplicateRecordDetector = new DuplicateRecordDetector();
+            let request: PotentialDuplicateRequest = new PotentialDuplicateRequest();
+            request.EntityID = this.EntityID;
+            request.ListID = this.SourceListID;
+            request.Options = {
+                DuplicateRunID: this.ID,
+            };
+            
+            const response = await duplicateRecordDetector.getDuplicateRecords(request, this.ContextCurrentUser);
+        }
+
+        return saveResult;
+    }
+}