+++ packages/AI/Core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log - @memberjunction/ai
 
+## 5.2.0
+
+### Patch Changes
+
+- @memberjunction/global@5.2.0
+
 ## 5.1.0
 
 ### Patch Changes