+++ packages/Angular/Generic/export-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @memberjunction/ng-export-service
 
+## 5.2.0
+
+### Patch Changes
+
+- @memberjunction/export-engine@5.2.0
+
 ## 5.1.0
 
 ### Patch Changes